Implemented item quantity addition
implements a concise function add_quantity that allows the user to add quantity for a selected item. It prompts the user for the item code, looks up the item ID, and if found, prompts for the quantity to add. Finally, it calls the add_quantity method from the DB object or module, passing the item ID and the new quantity as arguments.